To fix bug with event groups disappearing from the filter dropdown list after an event has been selected from the event_list template and the event_list template is returned to via the Back button on the next breadcrumb.
This happend on the "Three" journey due to the
scope.event_group_manually_set Bool being assigned a value of true when
scope.event_group_manually_set === null and scope.current_item.event_group
!= null.

In the "Three" journey, the scope.event_group_manually_set was assigned a
value of false as expected when the EventList controller initialised the
first time, but then it would somewhow lose this value when returning to
the event_list template and become 'undefined', and because
scope.current_item.event_group IS an object at this point,
scope.event_group_manually_set is then assigned a value of true.

As a result of this the setEventGroup method is called on the
$scope.current_item on line 96:

$scope.current_item.setEventGroup(event_group) if
$scope.event_group_manually_set

which I am guessing told the widget that the event_group has been
selected, so there is no need to iterate through a list of event_groups
and display them in a dropdown list, or some such logic.

So I added the if condition on line 88 to prevent this from happening,
without intefering with the current process should
$scope.event_group_manually_set be set to true elsewhere.